refactor: 拆分本地背景同步逻辑hook和启动通知hook
